src/requirements.txt
PIL==1.1.7
django==1.4.3
south==0.7.6
python-dateutil
